requirements.txt
Django==4.2.9
djangorestframework
pyyaml
requests
django-cors-headers
algoliasearch-django>=3.0,<4.0
